// src/api/types/PlanetType.ts
interface PlanetType {
  imageUrl: string;
  name: string;
  descriptions: string;
  rotation_period: string;
  orbital_period: string;
  diameter: string;
  climate: string;
  gravity: string;
  terrain: string;
  surface_water: string;
  population: string;
  residents: string[];
  films: string[];
  created: Date;
  edited: Date;
  url?: string;
}

export default PlanetType;